requirements.txt
flask
pytest
bcrypt
flask_sqlalchemy
psycopg2-binary
boto3
pymysql
python-dotenv